frontend/requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
appwrite==16.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
packaging==26.0
pefile==2024.8.26
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==82.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
